Move ImgFormat to Orange.widgets package
